# Conflicts:
#	src/main/java/com/example/onlineeduplatformlecture/handler/LectureHandler.java
#	src/main/java/com/example/onlineeduplatformlecture/repository/LectureRepository.java
#	src/main/java/com/example/onlineeduplatformlecture/service/LectureService.java
#	src/main/java/com/example/onlineeduplatformlecture/service/impl/LectureServiceImpl.java

+++ src/main/java/com/example/onlineeduplatformlecture/service/ContentService.java
@@ -10,9 +10,9 @@ import reactor.core.publisher.Mono;
 public interface ContentService {
 
     // GET /lectures/{lectureId}/contents (콘텐츠 리스트 조회)
-    Flux<Content> getContentList(int  lectureId);
+    Flux<Content> getContentList(long  lectureId);
 
     // GET /lectures/{lectureId}/contents/{contentId} (콘텐츠 열람)
-    Mono<Content> getContent(int lectureId, int contentId);
+    Mono<Content> getContent(long lectureId, long contentId);
 
 }
